requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
psycopg2-binary
python-jose
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-exporter-otlp 
python-json-logger
colorlog